requirements.txt
fastapi==0.86.0
uvicorn==0.19.0
pandas==1.3.5
newsapi-python===0.2.6
